+++ Cql/CoreTests/PrimitiveTests.cs
@@ -13,8 +13,8 @@ using System.ComponentModel;
 using System.IO;
 using System.Linq;
 using FluentAssertions;
-using Hl7.Cql.CodeGeneration.NET.Hosting;
-using Hl7.Cql.Compiler.Hosting;
+using Hl7.Cql.CodeGeneration.NET.Services;
+using Hl7.Cql.Compiler.Services;
 using DateTimePrecision = Hl7.Cql.Iso8601.DateTimePrecision;
 using Expression = System.Linq.Expressions.Expression;
 using Microsoft.Extensions.DependencyInjection;
