+++ Cql/CqlToElmTests/InvocationBuilderTests.cs
@@ -4,7 +4,7 @@ using Hl7.Cql.Elm;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Linq;
-using Hl7.Cql.CqlToElm.Hosting;
+using Hl7.Cql.CqlToElm.Services;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Hl7.Cql.CqlToElm.Test
